Fixed unittest with prefix and suffix
