+++ anyblok/tests/conftest.py
@@ -142,3 +142,21 @@ def registry_testblok_func(request, testbloks_loaded):
     registry = init_registry_with_bloks([], None)
     request.addfinalizer(registry.close)
     return registry
+
+
+@pytest.fixture(
+    scope="class",
+    params=[
+        ('prefix', 'suffix'),
+        ('', ''),
+    ]
+)
+def db_schema(request, bloks_loaded):
+    Configuration.set('prefix_db_schema.Model.*', request.param[0])
+    Configuration.set('suffix_db_schema.Model.*', request.param[1])
+
+    def rollback():
+        Configuration.set('prefix_db_schema.Model.*', '')
+        Configuration.set('suffix_db_schema.Model.*', '')
+
+    request.addfinalizer(rollback)
